pyproject.toml
[project]
name = "kws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "ffmpeg>=1.4",
    "hydra-core>=1.3.2",
    "netron>=8.7.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pytorch-lightning>=2.6.0",
    "six>=1.17.0",
    "tensorboard>=2.20.0",
    "thop>=0.1.1.post2209072238",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "torchmetrics>=1.8.2",
]
